backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain-mcp-m2m
langchain-anthropic
langgraph


